iommu: Make decision about needing IOMMU for hardware domains in advance
The hardware domains require IOMMU to be used in the most cases and
a decision to use it is made at hardware domain construction time.
But, it is not the best moment for the non-shared IOMMUs due to
the necessity of retrieving the domain memory page by page and re-creating
IOMMU mappings for the domain memory to eventually reach IOMMU page tables.

Make a decision about needing IOMMU a bit earlier, at domain creation time.
Having "d->need_iommu" flag set at the early stage we won't skip any
IOMMU mapping updates. So, when pages are added to the P2M they will be also
added to the IOMMU page tables if required. And as the result the existing
in iommu_hwdom_init() code that goes through the list of the domain pages
and tries to re-create IOMMU mappings won't be needed anymore and can be
just dropped.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>
CC: Jan Beulich <jbeulich@suse.com>
CC: Julien Grall <julien.grall@arm.com>
CC: Roger Pau Monne <roger.pau@citrix.com>